feat(uport): allow requests to be cancelled and qrdisplay to cancel
This returns a function that can be called to cancel a request, a request polls untils it gets a

response, but if a user cancels on their phone then they must be able to be given a way to cancel it

in the app, tthis allows developers to provide that.  A future alternative would be for the phone to

return a cancel response.